requirements.txt
zope.interface>=4.5.0
zope.configuration>=4.1.0
zope.component>=4.4.1
requests>=2.18.4
zulu>=0.12.0
attrs>=18.1.0
babel>=2.5.3
